+++ lib/src/cli/manual_command.dart
@@ -36,7 +36,7 @@ class ManualCommand extends Command<int> {
 
   @override
   Future<int> run() async {
-    final output = argResults!['output'] as String;
+    final output = argResults!.option('output')!;
     if (output == '-') {
       DartricsIO.stdoutSink.write(manualText);
     } else {